removeBlacklistedVariables moved to a separate file
