+++ apps/webapp/app/v3/environmentVariables/environmentVariablesRepository.server.ts
@@ -15,6 +15,7 @@ import {
   type Repository,
   type Result,
 } from "./repository";
+import { removeBlacklistedVariables } from "../environmentVariableRules.server";
 
 function secretKeyProjectPrefix(projectId: string) {
   return `environmentvariable:${projectId}:`;
@@ -1037,43 +1038,4 @@ async function resolveCommonBuiltInVariables(
   runtimeEnvironment: RuntimeEnvironmentForEnvRepo
 ): Promise<Array<EnvironmentVariable>> {
   return [];
-}
-
-type VariableRule =
-  | { type: "exact"; key: string }
-  | { type: "prefix"; prefix: string }
-  | { type: "whitelist"; key: string };
-
-const blacklistedVariables: VariableRule[] = [
-  { type: "exact", key: "TRIGGER_SECRET_KEY" },
-  { type: "exact", key: "TRIGGER_API_URL" },
-  { type: "prefix", prefix: "OTEL_" },
-  { type: "whitelist", key: "OTEL_LOG_LEVEL" },
-];
-
-export function removeBlacklistedVariables(
-  variables: EnvironmentVariable[]
-): EnvironmentVariable[] {
-  return variables.filter((v) => {
-    const whitelisted = blacklistedVariables.find(
-      (bv) => bv.type === "whitelist" && bv.key === v.key
-    );
-    if (whitelisted) {
-      return true;
-    }
-
-    const exact = blacklistedVariables.find((bv) => bv.type === "exact" && bv.key === v.key);
-    if (exact) {
-      return false;
-    }
-
-    const prefix = blacklistedVariables.find(
-      (bv) => bv.type === "prefix" && v.key.startsWith(bv.prefix)
-    );
-    if (prefix) {
-      return false;
-    }
-
-    return true;
-  });
 }
